+++ include/utils/logging.h
@@ -12,6 +12,7 @@ void serialLog(std::string value, int serialLoggingLevel);
 void serialLogln(const char *message, int serialLoggingLevel);
 void serialLogln(int value, int serialLoggingLevel);
 void serialLogln(double value, int serialLoggingLevel);
+void serialLogln(float value, int serialLoggingLevel);
 void serialLogln(std::string value, int serialLoggingLevel);
 
 void serialLogError(char message[], int error);
