+++ src/main.cpp
@@ -16,10 +16,12 @@
 #include "robot/encoder.h"
 #include "../env.h"
 #include "robot/pidController.h"
+#include "robot/magnet.h"
 
 //alright SCREW YOU serial monitor i won't print every frame then if you wanna play that game
 const int8_t PRINT_INTERVAL = 60;
 int8_t framesUntilPrint = 60;
+Magnet magnet;
 
 // Setup gets run at startup
 void setup() {
@@ -69,6 +71,9 @@ void loop() {
 
     // Run control loop
     controlLoop(loopDelayMilliseconds, framesUntilPrint);
+    MagnetReading mag_data = magnet.read_calibrated_data();
+    float heading = magnet.getCompassDegree(mag_data);
+    serialLogln(heading, 1);
 
     // This delay determines how often the code in loop is run
     // (Forcefully pauses the thread for about the amount of milliseconds passed in)
